+++ Automate/Framework/Machines/Buildings/JunimoHutMachine.cs
@@ -18,6 +18,9 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Buildings
         /// <summary>Whether fertilizer should be ignored when selecting output.</summary>
         private readonly bool IgnoreFertilizerOutput;
 
+        /// <summary>Whether to pull gemstones out of Junimo huts.</summary>
+        public bool PullGemstonesFromJunimoHuts { get; set; }
+
         /// <summary>The Junimo hut's output chest.</summary>
         private Chest Output => this.Machine.output.Value;
 
@@ -30,11 +33,13 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Buildings
         /// <param name="location">The location which contains the machine.</param>
         /// <param name="ignoreSeedOutput">Whether seeds should be ignored when selecting output.</param>
         /// <param name="ignoreFertilizerOutput">Whether fertilizer should be ignored when selecting output.</param>
-        public JunimoHutMachine(JunimoHut hut, GameLocation location, bool ignoreSeedOutput, bool ignoreFertilizerOutput)
+        /// <param name="pullGemstonesFromJunimoHuts">Whether to pull gemstones out of Junimo huts.</param>
+        public JunimoHutMachine(JunimoHut hut, GameLocation location, bool ignoreSeedOutput, bool ignoreFertilizerOutput, bool pullGemstonesFromJunimoHuts)
             : base(hut, location, BaseMachine.GetTileAreaFor(hut))
         {
             this.IgnoreSeedOutput = ignoreSeedOutput;
             this.IgnoreFertilizerOutput = ignoreFertilizerOutput;
+            this.PullGemstonesFromJunimoHuts = pullGemstonesFromJunimoHuts;
         }
 
         /// <summary>Get the machine's processing state.</summary>
@@ -77,7 +82,7 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Buildings
             foreach (Item item in this.Output.items.Where(p => p != null))
             {
                 // ignore gems which change Junimo colors (see JunimoHut:getGemColor)
-                if (item.Category == SObject.GemCategory || item.Category == SObject.mineralsCategory)
+                if (!this.PullGemstonesFromJunimoHuts && (item.Category == SObject.GemCategory || item.Category == SObject.mineralsCategory))
                     continue;
 
                 // ignore items used by another mod